feat: Long형을 int형으로 바꿈

+++ src/main/java/com/gnagnoohc/travel/community/dto/CommentDto.java
@@ -10,14 +10,14 @@ import lombok.Data;
 @Data
 @Alias("comment")
 public class CommentDto {
-	private Long commentId;           // comment_id  BIGINT  PK, AUTO_INCREMENT
+	private int commentId;            // comment_id  INT      PK, AUTO_INCREMENT
     private String content;           // content     TEXT    NOT NULL
-    private Integer depth;            // depth       INT     DEFAULT 0 (0=원댓글, 1=대댓글)
+    private int depth;                // depth       INT     DEFAULT 0 (0=원댓글, 1=대댓글)
     private Timestamp createdAt;  // created_at  DATETIME
     private Timestamp updatedAt;  // updated_at  DATETIME ON UPDATE
-    private Long parentId;            // parent_id   BIGINT  NULL (부모 댓글, 대댓글일 때만)
-    private Long postId;              // post_id     BIGINT  NOT NULL (FK → post)
-    private Long memberId;            // member_id   BIGINT  NOT NULL (FK → member)
+    private Integer parentId;         // parent_id   INT      NULL (부모 댓글일 때만 값 있음, top-level은 null → int 대신 Integer 유지)
+    private int postId;               // post_id     INT      NOT NULL (FK → post)
+    private int memberId;             // member_id   INT      NOT NULL (FK → member)
  
     // ── 화면 표시용 (테이블에 없는 조인/조립 값) ──
     private String memberName;        // 작성자 이름 (member 테이블 JOIN)
